Beta 0.0.3 and 0.1.0
Beta 0.0.3 adds two motors to the drive train (one on each side), and removed the auto-centergoal at the beginning.

Beta 0.1.0 is the beginning of the testing phase for the autonomous of the big robot. We will be using it to map out rotations and distances in order to carefully plan out our autonomous